Extreme performace improvements Now, cookies, headers, query strings, etc. are only parsed when they are used. After they are used once, they are cached for the rest of the request and then collected by the garbage collector.
This is great - if you don't need to parse the query string half the time, it won't get parsed half the time. It used to be that it was parsed for *every single request*.

Also, half the time that you do use it, and let's say you use it 200x per request (when it is used), it gets cached and is not re-parsed when you use it the second time for a request.